requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
nest-asyncio==1.5.6
python-dotenv==0.21.0
urllib3==1.26.13

